Changed version of variation-commons dependency. Moved mongo converter config to application configuration.
