+++ eva-server/src/main/java/uk/ac/ebi/eva/server/ApplicationConfiguration.java
@@ -17,6 +17,13 @@ package uk.ac.ebi.eva.server;
 
 import com.mongodb.MongoClient;
 
+import org.springframework.core.convert.converter.Converter;
+import org.springframework.data.mongodb.core.convert.CustomConversions;
+import org.springframework.data.mongodb.core.convert.DbRefResolver;
+import org.springframework.data.mongodb.core.convert.DefaultDbRefResolver;
+import org.springframework.data.mongodb.core.convert.MappingMongoConverter;
+import org.springframework.data.mongodb.core.mapping.MongoMappingContext;
+import uk.ac.ebi.eva.commons.models.converters.data.MongoDBObjectToVariantEntityConverter;
 import uk.ac.ebi.eva.lib.utils.DBAdaptorConnector;
 import uk.ac.ebi.eva.lib.utils.MultiMongoDbFactory;
 
@@ -25,6 +32,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.data.mongodb.MongoDbFactory;
 
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Properties;
 
 /**
@@ -45,4 +54,22 @@ public class ApplicationConfiguration {
         MongoClient mongoClient = DBAdaptorConnector.getMongoClient(properties);
         return new MultiMongoDbFactory(mongoClient, "test");
     }
+
+    @Bean
+    public CustomConversions customConversions() {
+        List<Converter<?, ?>> converters = new ArrayList<Converter<?, ?>>();
+        converters.add(new MongoDBObjectToVariantEntityConverter());
+        return new CustomConversions(converters);
+    }
+
+    @Bean
+    public MappingMongoConverter mongoConverter() throws IOException {
+        MongoMappingContext mappingContext = new MongoMappingContext();
+        DbRefResolver dbRefResolver = new DefaultDbRefResolver(mongoDbFactory());
+        MappingMongoConverter mongoConverter = new MappingMongoConverter(dbRefResolver, mappingContext);
+        mongoConverter.setCustomConversions(customConversions());
+        mongoConverter.afterPropertiesSet();
+        return mongoConverter;
+    }
+
 }
